fixed button, css, moved fetch create new post in services, added catch after each fetch in feedpage
