+++ src/app/components/Feed/CreatePostButton.js
@@ -11,9 +11,9 @@ export const CreatePostButton = (props) => {
             <i className="material-icons">add</i>
             </a>
             <ul>
-                <li><button data-target="modalPost" onClick = {props.handlerPostType} className="btn-floating"><i className="material-icons btn #4fc3f7 light-blue lighten-2  modal-trigger post">Text</i></button ></li>
-                <li><button data-target="modalImage" onClick = {props.handlerPostType} className="btn-floating"><i className="material-icons btn #1de9b6 teal accent-3 modal-trigger image">Image</i></button ></li>
-                <li><button data-target="modalVideo" onClick = {props.handlerPostType} className="btn-floating"><i className="material-icons btn #dd2c00 deep-orange accent-4 modal-trigger video">Video</i></button ></li>
+                <li><button data-target="modalPost" onClick = {props.handlerPostType} className="btn-floating"><i className="material-icons #4fc3f7 light-blue lighten-2  modal-trigger">short_text</i></button ></li>
+                <li><button data-target="modalImage" onClick = {props.handlerPostType} className="btn-floating"><i className="material-icons #1de9b6 teal accent-3 modal-trigger">insert_photo</i></button ></li>
+                <li><button data-target="modalVideo" onClick = {props.handlerPostType} className="btn-floating"><i className="material-icons #dd2c00 deep-orange accent-4 modal-trigger">video_library</i></button ></li>
             </ul>
         </div>
